# openapi.yaml
openapi: 3.0.3
info:
  title: Quản lý CLB — API
  description: >
    API của Telegram Mini App quản lý CLB (điểm danh, chi phí, sao kê).
    Tham khảo thiết kế tại thiet-ke-mini-app-clb.md.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://your-app.vercel.app
    description: Production (thay bằng NEXT_PUBLIC_APP_URL thật)
...
security:
  - cookieAuth: []
paths:
...
  /api/telegram/webhook:
    post:
      tags: [telegram]
      summary: Nhận update từ Telegram Bot API
      description: >
        Được Telegram gọi (đã cấu hình qua setWebhook). Xử lý message (/start, /getid),
        poll_answer (ghi Attendance), callback_query (data "paid:<statementId>" → xác nhận thanh toán).
      security:
        - telegramSecretToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Telegram Update object (message | poll_answer | callback_query)
      responses:
        "200":
          description: Đã xử lý (luôn trả 200 cho Telegram trừ khi sai secret token)
        "401":
          description: Sai X-Telegram-Bot-Api-Secret-Token
...
  /api/statements:
    get:
      tags: [statements]
      summary: Danh sách sao kê tháng (kèm tên member, mới nhất trước)
      responses:
        "200":
          description: Danh sách MonthlyStatement
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyStatement"
        "401":
          description: Chưa đăng nhập
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: clb_session
      description: JWT session cookie (httpOnly), cấp sau khi đăng nhập qua /api/auth/telegram.
    telegramSecretToken:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Do Telegram tự gửi lên khi gọi webhook, khớp với TELEGRAM_WEBHOOK_SECRET.
    cronBearer:
      type: http
      scheme: bearer
      description: "Header Authorization: Bearer <CRON_SECRET>, do cron-job.org gửi lên."
  schemas:
...
    MonthlyStatement:
      type: object
      properties:
        _id:
          type: string
        member_id:
          type: string
        month:
          type: string
          example: "2026-07"
        total_sessions:
          type: integer
        total_amount:
          type: number
        status:
          type: string
          enum: [pending, paid_reported, approved]
        paid_reported_at:
          type: string
          format: date-time
        approved_at:
          type: string
          format: date-time
        approved_by:
          type: string
...